feat(web): fold the cutout tool into the NIRCam pages
The cutout GUI shipped as a standalone /cutouts navbar page (#367); it now
lives where users browse the data it serves:

- New field-scoped route /nircam/[field]/cutouts — the same tool wearing the
  NIRCam chrome (breadcrumbs, field-switcher dropdown via a new `linkTo`
  prop), minus its own field <select> (the route owns the field). Shareable
  params become /nircam/<field>/cutouts?ra=&dec=&fov=&bands= with the same
  auto-generate-on-arrival behavior.
- Discovery: a "Cutouts" button in the field-detail header (only when the
  field has a field-composite FitsGL dataset) and a "Cutout at this position"
  entry in the map right-click menu (FitsGL surface only — Leaflet-only
  fields have no cutout source), which lands prefilled and auto-generates.
- /cutouts deleted outright (shipped yesterday, no external links) and the
  navbar entry removed.
- User-facing copy generalized: no more "FitsGL tile pyramid" jargon — a
  field without a cutout source now reads "This field's imaging hasn't been
  prepared for the cutout service." The RICE-quantization photometry caveat
  stays (scientifically meaningful), reworded off internal terms.

Render-verified against local Supabase: cosmos (dataset) shows the header
button + full tool with dataset bands; uds (no dataset) hides the button and
its cutouts route shows the plain-language empty state; navbar entry gone.
